+++ CHANGELOG.md
@@ -5,6 +5,47 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.7] - To be released
+
+### Added
+
+- Stubbing of an arbitrary option (nepalez)
+
+  ```yaml
+  ---
+  - object: Rails.application
+    chain:
+      - env
+    actions:
+      - return: production
+  ```
+
+### Changed
+
+- Partially defined options will satisfy an expectation (nepalez)
+
+  ```yaml
+  ---
+  - class: Payment
+    chain:
+      - call
+    arguments:
+      - 1
+      - :overdraft: true
+    actions:
+      - return: 3
+  ```
+  
+  This works even though the key `:notify` was not defined by the stub:
+
+  ```ruby
+  Payment.call 1, overdraft: true, notify: true
+  ```
+  
+  Notice, that these method works for key arguments only
+  (symbolized hash as the last argument).
+
+
 ## [0.0.6] - [2019-06-09]
 
 ### Added